+++ assets/src/edit-story/components/movable/index.js
@@ -1,0 +1,102 @@
+/**
+ * External dependencies
+ */
+import Moveable from 'react-moveable';
+
+/**
+ * WordPress dependencies
+ */
+import { useRef } from '@wordpress/element';
+
+/**
+ * Internal dependencies
+ */
+import { useStory } from '../../app';
+
+function Movable( {
+	selectedEl,
+	targetEl,
+} ) {
+	const {
+		state: { selectedElements },
+		actions: { setPropertiesOnSelectedElements },
+	} = useStory();
+
+	const moveable = useRef();
+
+	const frame = {
+		translate: [ 0, 0 ],
+		rotate: selectedEl.rotationAngle,
+	};
+
+	const setStyle = ( target ) => {
+		target.style.transform = `translate(${ frame.translate[ 0 ] }px, ${ frame.translate[ 1 ] }px) rotate(${ frame.rotate }deg)`;
+	};
+
+	const resetMoveable = ( target ) => {
+		frame.translate = [ 0, 0 ];
+		setStyle( target );
+		if ( moveable.current ) {
+			moveable.current.updateRect();
+		}
+	};
+
+	return (
+		<Moveable
+			ref={ moveable }
+			target={ targetEl }
+			draggable={ true }
+			resizable={ true }
+			rotatable={ true }
+			onDrag={ ( { target, beforeTranslate } ) => {
+				frame.translate = beforeTranslate;
+				setStyle( target );
+			} }
+			onDragStart={ ( { set } ) => {
+				set( frame.translate );
+			} }
+			onDragEnd={ ( { target } ) => {
+				setPropertiesOnSelectedElements( { x: selectedEl.x + frame.translate[ 0 ], y: selectedEl.y + frame.translate[ 1 ] } );
+				resetMoveable( target );
+			} }
+			onResizeStart={ ( { setOrigin, dragStart } ) => {
+				setOrigin( [ '%', '%' ] );
+				if ( dragStart ) {
+					dragStart.set( frame.translate );
+				}
+			} }
+			onResize={ ( { target, width, height, drag } ) => {
+				target.style.width = `${ width }px`;
+				target.style.height = `${ height }px`;
+				frame.translate = drag.beforeTranslate;
+				setStyle( target );
+			} }
+			onResizeEnd={ ( { target } ) => {
+				setPropertiesOnSelectedElements( {
+					width: parseInt( target.style.width ),
+					height: parseInt( target.style.height ),
+					x: selectedEl.x + frame.translate[ 0 ],
+					y: selectedEl.y + frame.translate[ 1 ],
+				} );
+				resetMoveable( target );
+			} }
+			onRotateStart={ ( { set } ) => {
+				set( frame.rotate );
+			} }
+			onRotate={ ( { target, beforeRotate } ) => {
+				frame.rotate = beforeRotate;
+				setStyle( target );
+			} }
+			onRotateEnd={ ( { target } ) => {
+				setPropertiesOnSelectedElements( { rotationAngle: frame.rotate } );
+				resetMoveable( target );
+			} }
+			origin={ false }
+			pinchable={ true }
+			keepRatio={ 'image' === selectedElements[ 0 ].type }
+			renderDirections={ 'image' === selectedElements[ 0 ].type ? [ 'nw', 'ne', 'sw', 'se' ] : [ 'n', 's', 'e', 'w', 'nw', 'ne', 'sw', 'se' ] }
+		/>
+	);
+}
+
+export default Movable;
